refactor: move uniqueblockSlotAcc to base class
